Move that VarlenEntry helper method out of pg_attribute.h. Fixed up Doxygen after today's refactors.
