+++ src/include/storage/storage_util.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <string>
 #include <unordered_map>
 #include <utility>
 #include <vector>
@@ -137,5 +138,22 @@ class StorageUtil {
    * @return vector of column ids
    */
   static std::vector<storage::col_id_t> ProjectionListAllColumns(const storage::BlockLayout &layout);
+
+  /**
+   * Helper method to turn a string into a VarlenEntry
+   * @param str input to be turned into a VarlenEntry
+   * @return varlen entry representing string
+   * @warning checking IsInlined() to see if you need to possibly clean up a buffer
+   */
+  static storage::VarlenEntry CreateVarlen(const std::string &str) {
+    if (str.size() > storage::VarlenEntry::InlineThreshold()) {
+      byte *contents = common::AllocationUtil::AllocateAligned(str.size());
+      std::memcpy(contents, str.data(), str.size());
+      return storage::VarlenEntry::Create(contents, static_cast<uint32_t>(str.size()), true);
+    } else {
+      return storage::VarlenEntry::CreateInline(reinterpret_cast<const byte *>(str.data()),
+                                                static_cast<uint32_t>(str.size()));
+    }
+  }
 };
 }  // namespace terrier::storage
